Clean up lib
