+++ crates/polars-io/src/avro/write.rs
@@ -1,3 +1,5 @@
+use std::io::Write;
+
 pub use arrow::io::avro::avro_schema::file::Compression;
 use arrow::io::avro::avro_schema::{self};
 use arrow::io::avro::write;
@@ -5,6 +7,7 @@ use polars_core::error::to_compute_err;
 pub use Compression as AvroCompression;
 
 use super::*;
+use crate::shared::SerWriter;
 
 /// Write a [`DataFrame`] to [Apache Avro] format
 ///
