+++ crates/polars-io/src/prelude.rs
@@ -1,6 +1,7 @@
 #[cfg(test)]
 use polars_core::prelude::*;
 
+pub use crate::cloud;
 #[cfg(feature = "csv")]
 pub use crate::csv::{read::*, write::*};
 #[cfg(any(feature = "ipc", feature = "ipc_streaming"))]
@@ -11,8 +12,8 @@ pub use crate::json::*;
 pub use crate::ndjson::core::*;
 #[cfg(feature = "parquet")]
 pub use crate::parquet::{metadata::*, read::*, write::*};
+pub use crate::shared::{SerReader, SerWriter};
 pub use crate::utils::*;
-pub use crate::{cloud, SerReader, SerWriter};
 
 #[cfg(test)]
 pub(crate) fn create_df() -> DataFrame {
